[common, shortcutguide] move d2d files from common to scg
