+++ src/common/pch.h
@@ -2,12 +2,7 @@
 #include <winrt/base.h>
 #include <winrt/Windows.Foundation.Collections.h>
 #include <Windows.h>
-#include <dxgi1_3.h>
-#include <d3d11_2.h>
-#include <d2d1_3.h>
-#include <d2d1_3helper.h>
-#include <d2d1helper.h>
-#include <dwrite.h>
+
 #include <dcomp.h>
 #include <dwmapi.h>
 #include <Shobjidl.h>
