+++ src/modules/shortcut_guide/d2d_svg.cpp
