+++ src/modules/shortcut_guide/d2d_text.h
