+++ src/modules/shortcut_guide/pch.h
@@ -1,4 +1,5 @@
 #pragma once
+#define NOMINMAX
 #include <winrt/base.h>
 #include <winrt/Windows.Foundation.h>
 #include <winrt/Windows.Foundation.Collections.h>
