+++ src/modules/shortcut_guide/shortcut_guide.vcxproj
@@ -105,6 +105,9 @@
     </ClCompile>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClInclude Include="d2d_svg.h" />
+    <ClInclude Include="d2d_text.h" />
+    <ClInclude Include="d2d_window.h" />
     <ClInclude Include="overlay_window.h" />
     <ClInclude Include="keyboard_state.h" />
     <ClInclude Include="Generated Files/resource.h" />
@@ -116,6 +119,9 @@
     <ClInclude Include="trace.h" />
   </ItemGroup>
   <ItemGroup>
+    <ClCompile Include="d2d_svg.cpp" />
+    <ClCompile Include="d2d_text.cpp" />
+    <ClCompile Include="d2d_window.cpp" />
     <ClCompile Include="overlay_window.cpp" />
     <ClCompile Include="dllmain.cpp" />
     <ClCompile Include="keyboard_state.cpp" />
